+++ karma.conf.js
@@ -1,22 +1,33 @@
 process.env.CHROME_BIN = require('puppeteer').executablePath();
 
-function testFiles(pkg) {
-  let files = [
-    'test/spec/features/**/*spec.js',
-    'test/spec/spec-helper.js',
-    'test/spec/util-spec.js',
-    'test/spec/cloudinary-spec.js',
-    'test/spec/transformation-spec.js',
-    'test/spec/tagspec.js',
-    'test/spec/videourlspec.js',
-    'test/spec/chaining-spec.js',
-    'test/spec/layer-spec.js',
-    'test/spec/lazy-load-spec.js',
-    `test/spec/responsive-${pkg}-spec.js`,
-  ];
-  if (pkg === 'jquery-file-upload') {
-    files.push('test/spec/cloudinary-jquery-upload-spec.js');
+let fs = require('fs');
+let glob = require('glob');
+
+function returnFilePathOrThrow(filePath) {
+  if (!fs.existsSync(filePath)) {
+    throw `Cannot find ${filePath}`;
+  } else {
+    return filePath;
   }
+}
+
+function testFiles(pkg) {
+  // all files under test/specspec/tests/
+  const files = glob.sync('test/spec/automatic/**/*.js'  , { cwd:process.cwd() });
+
+  if (files.length === 0) {
+    throw 'Error in glob pattern, no tests found';
+  }
+
+  // Ensure the responsive file exists add it
+  files.push(returnFilePathOrThrow(`test/spec/manual/responsive-${pkg}-spec.js`));
+
+  if (pkg === 'jquery-file-upload') {
+    files.push(returnFilePathOrThrow('test/spec/manual/cloudinary-jquery-upload-spec.js'));
+  } else{
+    files.push(returnFilePathOrThrow(`test/spec/manual/lazy-load-${pkg}-spec.js`));
+  }
+
   return files;
 }
 
@@ -55,9 +66,31 @@ module.exports = function(config) {
   let {minified, pkg='core'} = config.cloudinary || {};
 
   console.log(`Testing ${minified ? 'minified' : 'un-minified'}`);
-  const subject = `dist/cloudinary-${pkg}${minified ? '.min' : ''}.js`;
+  const subject = returnFilePathOrThrow(`dist/cloudinary-${pkg}${minified ? '.min' : ''}.js`);
+  const lazyLoadBase = returnFilePathOrThrow(`test/spec/manual/lazyLoadTestBase.js`);
 
-  const responsiveHtmlFile = `test/docRoot/responsive-${pkg}-test.html`;
+  const responsiveHtmlFile = returnFilePathOrThrow(`test/docRoot/responsive-${pkg}-test.html`);
+
+  let lazyLoadHtmlFile;
+  if (pkg !== 'jquery-file-upload') {
+    lazyLoadHtmlFile = returnFilePathOrThrow(`test/docRoot/lazy-load-${pkg}-test.html`);
+  }
+
+  let pattern = [
+    '{', // keep this first
+    `dist/*.map`
+  ];
+
+  if (responsiveHtmlFile) {
+    pattern.push(responsiveHtmlFile);
+  }
+
+  pattern = pattern.concat([
+    lazyLoadHtmlFile,
+    `node_modules/bootstrap/dist/+(css|js)/*`,
+    `test/docRoot/css/logo-nav.css`,
+    '}' // keep this last
+  ]).join(',');
 
   return config.set({
     // base path that will be used to resolve all patterns (eg. files, exclude)
@@ -70,16 +103,10 @@ module.exports = function(config) {
     files: [
       ...dependency(pkg),
       subject,
+      lazyLoadBase,
       ...testFiles(pkg),
       {
-        pattern: [
-          '{', // keep this first
-          `dist/*.map`,
-          responsiveHtmlFile,
-          `node_modules/bootstrap/dist/+(css|js)/*`,
-          `test/docRoot/css/logo-nav.css`,
-          '}' // keep this last
-        ].join(','),
+        pattern,
         watched: false,
         included: false,
         served: true,