+++ src/cloudinary.js
@@ -17,7 +17,6 @@ import {
   assign,
   defaults,
   getData,
-  isArray,
   isEmpty,
   isFunction,
   isString,
@@ -486,25 +485,26 @@ class Cloudinary {
    * @see {@link Cloudinary#cloudinary_update|cloudinary_update} for additional configuration parameters
    * @see <a href="https://cloudinary.com/documentation/responsive_images#automating_responsive_images_with_javascript"
    *  target="_blank">Automating responsive images with JavaScript</a>
+   * @return {function} that when called, removes the resize EventListener added by this function
    */
   responsive(options, bootstrap = true) {
     var ref, ref1, ref2, responsiveClass, responsiveResize, timeout;
     this.responsiveConfig = merge(this.responsiveConfig || {}, options);
-    responsiveClass = (ref = this.responsiveConfig['responsive_class']) != null ? ref : this.config('responsive_class');
+    responsiveClass = (ref = this.responsiveConfig.responsive_class) != null ? ref : this.config('responsive_class');
     if (bootstrap) {
       this.cloudinary_update(`img.${responsiveClass}, img.cld-hidpi`, this.responsiveConfig);
     }
-    responsiveResize = (ref1 = (ref2 = this.responsiveConfig['responsive_resize']) != null ? ref2 : this.config('responsive_resize')) != null ? ref1 : true;
+    responsiveResize = (ref1 = (ref2 = this.responsiveConfig.responsive_resize) != null ? ref2 : this.config('responsive_resize')) != null ? ref1 : true;
     if (responsiveResize && !this.responsiveResizeInitialized) {
       this.responsiveConfig.resizing = this.responsiveResizeInitialized = true;
       timeout = null;
-      return window.addEventListener('resize', () => {
+      const makeResponsive = () => {
         var debounce, ref3, ref4, reset, run, wait, waitFunc;
-        debounce = (ref3 = (ref4 = this.responsiveConfig['responsive_debounce']) != null ? ref4 : this.config('responsive_debounce')) != null ? ref3 : 100;
+        debounce = (ref3 = (ref4 = this.responsiveConfig.responsive_debounce) != null ? ref4 : this.config('responsive_debounce')) != null ? ref3 : 100;
         reset = function() {
           if (timeout) {
             clearTimeout(timeout);
-            return timeout = null;
+            timeout = null;
           }
         };
         run = () => {
@@ -516,14 +516,16 @@ class Cloudinary {
         };
         wait = function() {
           reset();
-          return timeout = setTimeout(waitFunc, debounce);
+          timeout = setTimeout(waitFunc, debounce);
         };
         if (debounce) {
           return wait();
         } else {
           return run();
         }
-      });
+      };
+      window.addEventListener('resize', makeResponsive);
+      return ()=>window.removeEventListener('resize', makeResponsive);
     }
   }
 
@@ -684,11 +686,14 @@ class Cloudinary {
               setUrl = false;
             }
           }
-          if(options.loading === 'lazy' && !this.isNativeLazyLoadSupported() && this.isLazyLoadSupported() && !elements[0].getAttribute('src')) {
-            this.setImgOnLazyLoad(elements, options);
-          }else if (setUrl) {
+          const isLazyLoading = (options.loading === 'lazy' && !this.isNativeLazyLoadSupported() && this.isLazyLoadSupported() && !elements[0].getAttribute('src'));
+          if (setUrl || isLazyLoading){
+            // If data-width exists, set width to be data-width
+            this.setAttributeIfExists(elements[0], 'width', 'data-width');
+          }
+
+          if (setUrl && !isLazyLoading) {
             setAttribute(tag, 'src', dataSrc);
-            elements[0].setAttribute('width', elements[0].getAttribute('data-width'));
           }
         }
       }
@@ -697,12 +702,16 @@ class Cloudinary {
   }
 
   /**
-   * Sets width when not using native lazy load
-   * @param img
-   * @param options
+   * Sets element[toAttribute] = element[fromAttribute] if element[fromAttribute] is set
+   * @param element
+   * @param toAttribute
+   * @param fromAttribute
    */
-  setImgOnLazyLoad(img, options){
-    img[0].setAttribute('width', img[0].getAttribute('data-width'));
+  setAttributeIfExists(element, toAttribute, fromAttribute){
+    const attributeValue = element.getAttribute(fromAttribute);
+    if (attributeValue != null) {
+      setAttribute(element, toAttribute, attributeValue);
+    }
   }
 
   /**