+++ src/configuration.js
@@ -25,11 +25,11 @@ class Configuration {
   }
 
   /**
-   * Initializes the configuration. This method is a convenience method that invokes both 
-   *  {@link Configuration#fromEnvironment|fromEnvironment()} (Node.js environment only) 
-   *  and {@link Configuration#fromDocument|fromDocument()}. 
-   *  It first tries to retrieve the configuration from the environment variable. 
-   *  If not available, it tries from the document meta tags. 
+   * Initializes the configuration. This method is a convenience method that invokes both
+   *  {@link Configuration#fromEnvironment|fromEnvironment()} (Node.js environment only)
+   *  and {@link Configuration#fromDocument|fromDocument()}.
+   *  It first tries to retrieve the configuration from the environment variable.
+   *  If not available, it tries from the document meta tags.
    * @function Configuration#init
    * @return {Configuration} returns `this` for chaining
    * @see fromDocument