+++ src/transformation.js
@@ -1,6 +1,7 @@
 import Expression from './expression';
 import Condition from './condition';
 import Configuration from './configuration';
+import {URL_KEYS} from './constants';
 
 import {
   assign,
@@ -408,11 +409,13 @@ class TransformationBase {
    * @return PlainObject
    */
   toHtmlAttributes() {
-    var attrName, height, options, ref2, ref3, value, width;
+    let attrName, height, options, ref2, ref3, value, width;
     options = {};
+    let snakeCaseKey;
     Object.keys(this.otherOptions).forEach(key=>{
       value = this.otherOptions[key];
-      if (!contains(Transformation.PARAM_NAMES, snakeCase(key))) {
+      snakeCaseKey = snakeCase(key);
+      if (!contains(Transformation.PARAM_NAMES, snakeCaseKey) && !contains(URL_KEYS, snakeCaseKey)) {
         attrName = /^html_/.test(key) ? key.slice(5) : key;
         options[attrName] = value;
       }
@@ -755,6 +758,10 @@ class Transformation extends TransformationBase {
     return this.param(value, "keyframe_interval", "ki");
   }
 
+  ocr(value) {
+    return this.param(value, "ocr", "ocr");
+  }
+
   offset(value) {
     var end_o, start_o;
     [start_o, end_o] = (isFunction(value != null ? value.split : void 0)) ? value.split('..') : isArray(value) ? value : [null, null];
@@ -909,6 +916,7 @@ Transformation.methods = [
   "htmlWidth",
   "if",
   "keyframeInterval",
+  "ocr",
   "offset",
   "opacity",
   "overlay",