+++ src/util/lodash.js
@@ -1,5 +1,9 @@
 var nodeContains;
 
+import getSDKAnalyticsSignature from "../sdkAnalytics/getSDKAnalyticsSignature";
+import getAnalyticsOptions from "../sdkAnalytics/getAnalyticsOptions";
+export {getSDKAnalyticsSignature , getAnalyticsOptions};
+
 export {
   default as assign
 } from 'lodash/assign';
@@ -52,11 +56,12 @@ import isElement from 'lodash/isElement';
 import isFunction from 'lodash/isFunction';
 import trim from 'lodash/trim';
 
+export * from './lazyLoad';
 export * from './baseutil';
 export {
   isElement,
   isFunction,
-  trim}
+  trim};
 /*
  * Includes utility methods and lodash / jQuery shims
  */