+++ types/cloudinary-core.d.ts
@@ -22,9 +22,28 @@ type VideoFlags = string | Array<string> | "splice" | "layer_apply" | "no_stream
 type AudioCodec = string | "none" | "aac" | "vorbis" | "mp3";
 type AudioFrequency = number | 8000 | 11025 | 16000 | 22050 | 32000 | 37800 | 44056 | 44100 | 47250 | 48000 | 88200 | 96000 | 176400 | 192000;
 type StreamingProfiles = string | "4k" | "full_hd" | "hd" | "sd" | "full_hd_wifi" | "full_hd_lean" | "hd_lean";
+type UrlOptions = (Transformation | Transformation.Options) & { placeholder?: string, accessibility?: string };
 
 export function crc32(str: string): any;
 export function utf8_encode(argString: string): any;
+
+type AnalyticsOptions = {
+    sdkSemver: string;
+    techVersion: string;
+    sdkCode: string;
+    feature: string;
+}
+
+type AnalyticsOptionsParameters = {
+    sdkSemver: string;
+    techVersion: string;
+    sdkCode: string;
+    urlAnalytics?: boolean;
+    accessibility?: boolean;
+    loading?: string;
+    responsive?: boolean;
+    placeholder?: boolean;
+}
 
 export class Util {
     static allStrings(list: Array<any>): boolean;
@@ -60,6 +79,10 @@ export class Util {
     static identity(value: any): any;
     static isPlainObject(value: any): boolean;
     static trim(text: string): string;
+    static detectIntersection(element: Element, onIntersect: Function): void
+
+    static getAnalyticsOptions(options: AnalyticsOptionsParameters) : AnalyticsOptions;
+    static getSDKAnalyticsSignature(options: AnalyticsOptions):string;;
 }
 
 
@@ -550,7 +573,7 @@ export class TextLayer extends Layer {
     fontWeight(value: string): TextLayer;
     fontStyle(value: string): TextLayer;
     fontHinting(value: string): TextLayer;
-    fontAntiAliasing(value: string): TextLayer;
+    fontAntialiasing(value: string): TextLayer;
     textDecoration(value: string): TextLayer;
     textAlign(value: string): TextLayer;
     stroke(value: string): TextLayer;
@@ -630,7 +653,11 @@ export class Param {
  * @param {Object} options - options to configure Cloudinary
  * @see Configuration for more details
  * @example
+ *    // Include cloudinary_js in a <script> tag, then:
  *    var cl = new cloudinary.Cloudinary( { cloud_name: "mycloud"});
+ *    // Or import it from the npm package:
+ *    import { Cloudinary } from 'cloudinary-core';
+ *    var cl = new Cloudinary( { cloud_name: "mycloud"});
  *    var imgTag = cl.image("myPicID");
  */
 export class Cloudinary {
@@ -642,12 +669,12 @@ export class Cloudinary {
      * @function Cloudinary#url
      * @param {string} publicId - the public ID of the resource
      * @param {Object} [options] - options for the tag and transformations, possible values include all {@link Transformation} parameters
-     *                          and {@link Configuration} parameters
+     *                          and {@link Configuration} parameters with addition of 'placeholder' and 'accessibility'
      * @param {string} [options.type='upload'] - the classification of the resource
      * @param {Object} [options.resource_type='image'] - the type of the resource
      * @return {string} The resource URL
      */
-    url(publicId: string, options?: Transformation | Transformation.Options): string;
+    url(publicId: string, options?: UrlOptions ): string;
 
     /**
      * Generate an video resource URL.